+++ docs/changelog.md
@@ -2,6 +2,7 @@
 - Replaces data backend with DuckDB + Ibis, making plotting 2-3x faster and object lookup up to 10x faster
 - Changes default unit of right ascension to degrees (0...360)
 - Improves performance of label placement
+- Separates the constellation Serpens into two "separate" parts (Cauda and Caput)
 
 ## v0.14.x
 [Documentation](https://archives.starplot.dev/0.14.0/)